+++ khFinalproject/src/main/java/com/kh/spring/controller/HomeController.java
@@ -1,15 +1,8 @@
 package com.kh.spring.controller;
 
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-
 
 import lombok.extern.slf4j.Slf4j;
